Added services for payments

+++ CRS.WebApi/Data/TaxStatementResponse.cs
@@ -5,6 +5,6 @@ namespace CRS.WebApi.Data
     public class TaxStatementResponse 
     {
         [JsonPropertyName("amountOwing")]
-        public int AmountOwing {  get; set; }
+        public decimal AmountOwing { get; set; }
     }
 }
